time sheet

+++ fha_subvention/__init__.py
@@ -2,30 +2,30 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import models
+from odoo.exceptions import ValidationError
 
 from odoo import api, SUPERUSER_ID
 
 import logging
+
 _logger = logging.getLogger(__name__)
+
 
 def uninstall_hook(cr, registry):
     env = api.Environment(cr, SUPERUSER_ID, {})
 
     try:
         accounts = env['account.analytic.account'].search([('subvention', '=', True)])
-        accounts.write({'active': False})
+        if accounts:
+            _logger.warning(
+                "The following subventions items have been archived following 'subventions' module uninstallation: %s" % accounts.ids)
+            accounts.unlink()
 
-        _logger.warning("The following subventions items have been archived following 'subventions' module uninstallation: %s" % accounts.ids)
-
-        accounts.unlink()
-    except:
-        pass
-
-    try:
         groups = env['account.analytic.group'].search([('subvention', '=', True)])
-
-        groups.unlink()
-    except:
-        pass
-
-    _logger.warning("The following subventions have been deleted following 'subventions' module uninstallation: %s" % groups.ids)
+        if groups:
+            _logger.warning(
+                "The following subventions have been deleted following 'subventions' module uninstallation: %s" % groups.ids)
+            groups.unlink()
+    except Exception as e:
+        _logger.exception("Transaction post processing failed")
+        env.cr.rollback()
